backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
scapy==2.5.0
python-nmap==0.7.1
python-multipart==0.0.6